Create extract_values_from_json.py
This function takes in a JSON object and a list of keys that you want to extract from the object. It returns a dictionary containing the extracted key-value pairs.

Here’s an example of how you can use this function:

# Define the JSON object
json_obj = {
    'key1': 'value1',
    'key2': {
        'key3': 'value3',
        'key4': 'value4'
    },
    'key5': [
        {
            'key6': 'value6',
            'key7': 'value7'
        },
        {
            'key8': 'value8',
            'key9': 'value9'
        }
    ]
}

# Define the keys to extract
keys = ['key3', 'key7', 'key9']

# Call the extract_elements function
result = extract_elements(json_obj, keys)

# Print the result
print(result)

In this example, we define a JSON object and a list of keys that we want to extract from the object. Then, we call the extract_elements function and pass in these values as arguments. The function returns a dictionary containing the extracted key-value pairs.